# KARNATAKA/games/Karnataka-Heritage-Games/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type GameResultDifficulty = typeof GameResultDifficulty[keyof typeof GameResultDifficulty];


export const GameResultDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export type GameResultStats = {[key: string]: number};

export type GameResultResult = typeof GameResultResult[keyof typeof GameResultResult];


export const GameResultResult = {
  win: 'win',
  loss: 'loss',
  draw: 'draw',
} as const;

export interface GameResult {
  id: number;
  playerId: string;
  gameId: string;
  gameName: string;
  difficulty: GameResultDifficulty;
  score: number;
  accuracy: number;
  durationSeconds: number;
  totalMoves: number;
  validMoves: number;
  invalidMoves: number;
  stats: GameResultStats;
  result: GameResultResult;
  startedAt: string;
  completedAt: string;
}

export type GameResultInputDifficulty = typeof GameResultInputDifficulty[keyof typeof GameResultInputDifficulty];


export const GameResultInputDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export type GameResultInputStats = {[key: string]: number};

export type GameResultInputResult = typeof GameResultInputResult[keyof typeof GameResultInputResult];


export const GameResultInputResult = {
  win: 'win',
  loss: 'loss',
  draw: 'draw',
} as const;

export interface GameResultInput {
  /** @minLength 1 */
  playerId: string;
  /** @minLength 1 */
  gameId: string;
  /** @minLength 1 */
  gameName: string;
  difficulty: GameResultInputDifficulty;
  score: number;
  accuracy: number;
  durationSeconds: number;
  totalMoves: number;
  validMoves: number;
  invalidMoves: number;
  stats: GameResultInputStats;
  result: GameResultInputResult;
  startedAt: string;
  completedAt: string;
}

export type GameStatsGameProgress = {[key: string]: boolean};

export interface GameStatsByGame {
  bestScore: number;
  bestAccuracy: number;
  fastestTime: number;
  attempts: number;
  wins: number;
  recentResults: GameResult[];
}

export type GameStatsGameStats = {[key: string]: GameStatsByGame};

export interface GameStats {
  totalGames: number;
  completedGames: number;
  bestScore: number;
  bestAccuracy: number;
  fastestTime: number;
  attempts: number;
  wins: number;
  gameProgress: GameStatsGameProgress;
  gameStats: GameStatsGameStats;
  recentResults: GameResult[];
}

export type ListGameResultsParams = {
/**
 * @minLength 1
 */
playerId: string;
};

export type GetGameStatsParams = {
/**
 * @minLength 1
 */
playerId: string;
};