# KARNATAKA/games/Karnataka-Heritage-Games/lib/api-zod/src/generated/types/gameResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GameResultDifficulty } from './gameResultDifficulty';
import type { GameResultResult } from './gameResultResult';
import type { GameResultStats } from './gameResultStats';

export interface GameResult {
  id: number;
  playerId: string;
  gameId: string;
  gameName: string;
  difficulty: GameResultDifficulty;
  score: number;
  accuracy: number;
  durationSeconds: number;
  totalMoves: number;
  validMoves: number;
  invalidMoves: number;
  stats: GameResultStats;
  result: GameResultResult;
  startedAt: Date;
  completedAt: Date;
}