# KARNATAKA/games/Karnataka-Heritage-Games/lib/api-zod/src/generated/types/gameResultInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GameResultInputDifficulty } from './gameResultInputDifficulty';
import type { GameResultInputResult } from './gameResultInputResult';
import type { GameResultInputStats } from './gameResultInputStats';

export interface GameResultInput {
  /** @minLength 1 */
  playerId: string;
  /** @minLength 1 */
  gameId: string;
  /** @minLength 1 */
  gameName: string;
  difficulty: GameResultInputDifficulty;
  score: number;
  accuracy: number;
  durationSeconds: number;
  totalMoves: number;
  validMoves: number;
  invalidMoves: number;
  stats: GameResultInputStats;
  result: GameResultInputResult;
  startedAt: Date;
  completedAt: Date;
}