# KARNATAKA/games/Karnataka-Heritage-Games/lib/api-zod/src/generated/types/gameStats.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GameResult } from './gameResult';
import type { GameStatsGameProgress } from './gameStatsGameProgress';
import type { GameStatsGameStats } from './gameStatsGameStats';

export interface GameStats {
  totalGames: number;
  completedGames: number;
  bestScore: number;
  bestAccuracy: number;
  fastestTime: number;
  attempts: number;
  wins: number;
  gameProgress: GameStatsGameProgress;
  gameStats: GameStatsGameStats;
  recentResults: GameResult[];
}