# KARNATAKA/games/Karnataka-Heritage-Games/lib/api-zod/src/generated/types/gameStatsByGame.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GameResult } from './gameResult';

export interface GameStatsByGame {
  bestScore: number;
  bestAccuracy: number;
  fastestTime: number;
  attempts: number;
  wins: number;
  recentResults: GameResult[];
}